HAJ-1 :: adding audit logs for public methods in controllers and also log for exceptions

+++ shj-admin-portal-orm/src/main/java/com/elm/shj/admin/portal/orm/entity/JpaAuditLog.java
@@ -55,7 +55,7 @@ public class JpaAuditLog implements Serializable {
     private Date startTime;
 
     @Column(name = "processing_time", nullable = false)
-    private int processingTime;
+    private long processingTime;
 
     @Column(name = "http_status", nullable = false)
     private int httpStatus;
